requirement.txt
openai
python-dotenv
pandas
notebook
scikit-learn
google-genai